# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BabySummaryDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "birthDate": {
            "format": "date",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nickname": {
            "maxLength": 30,
            "minLength": 1,
            "type": "string"
          },
          "purgeAfter": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "sex": {
            "enum": [
              "MALE",
              "FEMALE"
            ],
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "familyId",
          "nickname",
          "birthDate",
          "sex",
          "status",
          "archivedAt",
          "purgeAfter",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "BatchUpdatePhotosRequestDto": {
        "properties": {
          "capturedOn": {
            "format": "date",
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "nullable": true,
            "type": "string"
          },
          "location": {
            "maxLength": 80,
            "nullable": true,
            "type": "string"
          },
          "photoIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "photoIds"
        ],
        "type": "object"
      },
...
      "CreatePhotoBatchRequestDto": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/PhotoFileDescriptorDto"
            },
            "maxItems": 20,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "type": "object"
      },
...
      "PhotoFileDescriptorDto": {
        "properties": {
          "capturedOn": {
            "format": "date",
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "maximum": 20971520,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "contentType",
          "sizeBytes",
          "capturedOn"
        ],
        "type": "object"
      },
      "PhotoListResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PhotoSummaryDto"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "PhotoManagementPageDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PhotoSummaryDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
      "PhotoPreviewDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "expiresAt"
        ],
        "type": "object"
      },
...
      "PhotoUploadBatchDto": {
        "properties": {
          "babyId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoSummaryDto"
            },
            "type": "array"
          },
          "uploadInstructions": {
            "items": {
              "$ref": "#/components/schemas/PhotoUploadInstructionDto"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "babyId",
          "createdAt",
          "photos"
        ],
        "type": "object"
      },
...
      "PhotoUploadInstructionDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "photoId": {
            "format": "uuid",
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "photoId",
          "url",
          "fields",
          "expiresAt"
        ],
        "type": "object"
      },
      "PublishPhotosRequestDto": {
        "properties": {
          "photoIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "photoIds"
        ],
        "type": "object"
      },
...
      "UpdateBabyRequestDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "type": "string"
          },
          "nickname": {
            "maxLength": 30,
            "minLength": 1,
            "type": "string"
          },
          "sex": {
            "enum": [
              "MALE",
              "FEMALE"
            ],
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "UpdatePhotoRequestDto": {
        "properties": {
          "capturedOn": {
            "format": "date",
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "nullable": true,
            "type": "string"
          },
          "location": {
            "maxLength": 80,
            "nullable": true,
            "type": "string"
          },
          "title": {
            "maxLength": 80,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "Public API contract for Baby Growth Gallery",
    "title": "Baby Growth Gallery API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/families/{familyId}/babies/{babyId}": {
      "get": {
        "operationId": "BabyController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BabySummaryDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "babies"
        ]
      },
      "patch": {
        "operationId": "BabyController_update",
        "parameters": [
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBabyRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BabySummaryDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "babies"
        ]
      }
    },
...
    "/api/v1/families/{familyId}/babies/{babyId}/photo-upload-batches": {
      "post": {
        "operationId": "PhotoController_createBatch",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhotoBatchRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadBatchDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photo-upload-batches/{batchId}": {
      "get": {
        "operationId": "PhotoController_getBatch",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadBatchDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photo-upload-batches/{batchId}/photos": {
      "patch": {
        "operationId": "PhotoController_batchUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdatePhotosRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoListResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photo-upload-batches/{batchId}/photos/{photoId}/complete": {
      "post": {
        "operationId": "PhotoController_complete",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photo-upload-batches/{batchId}/photos/{photoId}/reissue": {
      "post": {
        "operationId": "PhotoController_reissue",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadInstructionDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photo-upload-batches/{batchId}/publish": {
      "post": {
        "operationId": "PhotoController_publish",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishPhotosRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoListResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photos/{photoId}": {
      "delete": {
        "operationId": "PhotoController_discard",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      },
      "patch": {
        "operationId": "PhotoController_update",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhotoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoSummaryDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photos/{photoId}/preview": {
      "get": {
        "operationId": "PhotoController_preview",
        "parameters": [
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "enum": [
                "THUMBNAIL",
                "DISPLAY",
                "ARCHIVE"
              ],
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoPreviewDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photos/{photoId}/restore": {
      "post": {
        "operationId": "PhotoController_restore",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoSummaryDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photos/{photoId}/trash": {
      "post": {
        "operationId": "PhotoController_trash",
        "parameters": [
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoSummaryDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/families/{familyId}/babies/{babyId}/photos/manage": {
      "get": {
        "operationId": "PhotoController_manage",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {}
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {}
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {}
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "enum": [
                "mine",
                "family"
              ],
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "familyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "babyId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoManagementPageDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
...
  },
  "servers": [],
...
}